requirements.txt
APScheduler==3.10.4
msal==1.24.1
google-cloud-storage==2.14.0
requests==2.31.0
aiohttp==3.9.1
psutil==7.2.1

